Add files via upload
Write a program that inputs two grades separated by a space.  
If both grades are below a score of sixty then the program should output the message:
Student Failed:(
If both grades are greater than or equal to ninety five then the program should output the message:
Student Graduated with Honors:)
Otherwise the program should output the message:
Student Graduated!